+++ python/acl_anthology/collections/collection.py
@@ -198,6 +198,7 @@ class Collection(SlottedDict[Volume]):
         # If editors were given, we fill in their ID & add them to the index
         if volume.editors:
             for namespec in volume.editors:
+                namespec.normalize()
                 self.root.people.ingest_namespec(namespec)
             self.root.people._add_to_index(volume.editors, volume.full_id_tuple)